requirements.txt
arxivscraper>=0.0.5
pandas>=1.0.0
python-dateutil>=2.8.2
openai>=1.0.0
python-dotenv>=1.0.0
